Split native_library from webview.h
